Make a separate data frame for each well.
I realized that caching the wells and referencing the same data frame
from multiple places was a bad idea, because it prevents you from
processing the same well differently in the context of different
experiments.  For example, imagine wanting to normalize the controls in
terms of RFP for one experiment and GFP for another.